requirements.txt
# LLM APIs
anthropic
aider-chat
backoff
openai
# Viz
matplotlib
pypdf
pymupdf4llm
# Common Requirements
torch
numpy
transformers
datasets
tiktoken
wandb
tqdm
requests

Pillow=11.1.0
